intermediate commit

+++ static/exports/haskell/Tac2e770f2132aced749ec197385ff552.hs
@@ -1,1 +1,3 @@
-module Typeable.Tac2e770f2132aced749ec197385ff552 where
+module Typeable.Tac2e770f2132aced749ec197385ff552 (Int) where
+
+import Data.Int
